bot_notification/requirements.txt
aiogram
confluent_kafka
apscheduler
sqlalchemy
loguru 
requests
pyjwt
schedule
pybase64